requirements.txt
PyGithub
flask
python-dotenv
requests
rich
